MAINT: signal, _buld_utils: move tempita template dance to _build_utils
