+++ scipy/signal/setup.py
@@ -1,28 +1,7 @@
 from scipy._build_utils import numpy_nodepr_api
+from scipy._build_utils import tempita
 import os
 import sys
-
-def process_tempita(fromfile):
-    try:
-        try:
-            from Cython import Tempita as tempita
-        except ImportError:
-            import tempita
-    except ImportError:
-        raise Exception("Building requires tempita")
-
-    if not fromfile.endswith('c.in'):
-        raise ValueError("Unexpected extension: %s" % fromfile)
-
-    from_filename = tempita.Template.from_filename
-    template = from_filename(fromfile,
-                             encoding=sys.getdefaultencoding())    
-
-    content = template.substitute()
-
-    outfile = os.path.splitext(fromfile)[0]
-    with open(outfile, 'w') as f:
-        f.write(content)
 
 
 def configuration(parent_package='', top_path=None):
@@ -37,8 +16,8 @@ def configuration(parent_package='', top_path=None):
 
     # convert the *.c.in files : `lfilter.c.in -> lfilter.c` etc
     srcdir = os.path.join(os.getcwd(), 'scipy', 'signal')
-    process_tempita(os.path.join(srcdir, 'lfilter.c.in') )
-    process_tempita(os.path.join(srcdir, 'correlate_nd.c.in') )
+    tempita.process_tempita(os.path.join(srcdir, 'lfilter.c.in') )
+    tempita.process_tempita(os.path.join(srcdir, 'correlate_nd.c.in') )
 
     sigtools = config.add_extension('sigtools',
                          sources=['sigtoolsmodule.c', 'firfilter.c',
